requirements.txt
python-dotenv
pandas
numpy
pymongo>=4.0
certifi
dnspython
dill
mlflow
dagshub
pyaml
fastapi
uvicorn
scikit-learn


##-e .